Cast values to int64

+++ pkg/hintrunner/zero/zerohint_usort_test.go
@@ -94,9 +94,9 @@ func TestZeroHintUsort(t *testing.T) {
 				},
 				ctxInit: func(ctx *hinter.HintRunnerContext) {
 					ctx.ScopeManager.EnterScope(map[string]any{
-						"positions":   []int{8, 6, 4},
-						"current_pos": 2,
-						"last_pos":    1,
+						"positions":   []int64{8, 6, 4},
+						"current_pos": int64(2),
+						"last_pos":    int64(1),
 					})
 				},
 				makeHinter: func(ctx *hintTestContext) hinter.Hinter {
@@ -104,8 +104,8 @@ func TestZeroHintUsort(t *testing.T) {
 				},
 				check: func(t *testing.T, ctx *hintTestContext) {
 					allVarValueInScopeEquals(map[string]any{
-						"current_pos": 4,
-						"last_pos":    3,
+						"current_pos": int64(4),
+						"last_pos":    int64(3),
 					})(t, ctx)
 
 					varValueEquals("next_item_index", feltInt64(1))(t, ctx)
